Move LayerKey into dygma.keys
